refactor(transport): regroup shared/ into internal/ subtrees by ownership
按真实归属把 transport/shared/* 5 个子目录搬到三处 internal/：

- transport/internal/normalize/         (双向: NormalizeDuration/NormalizeEndpoint)
- transport/internal/tls/               (双向: BuildServerTLS/BuildClientTLS)
- transport/server/internal/registry/   (server-only: ResolveRegistryEndpoint)
- transport/server/internal/accept/     (server-only: AcceptLoop)
- transport/client/internal/endpointindex/ (client-only: BuildClientEndpointIndex)

包名按主题更新（config/endpoint/lifecycle → normalize/registry/accept/endpointindex）。
4 个 consumer（server/grpc, server/http, client/grpc, client/http）的 import 同步更新。
transport/shared/ 目录删除；internal/ 机制阻止外部 import。

412 tests / 61 packages 通过。
