+++ transport/client/internal/endpointindex/endpoint_index.go
@@ -1,4 +1,4 @@
-package config
+package endpointindex
 
 import (
 	"fmt"
