+++ transport/client/internal/endpointindex/endpoint_index_test.go
@@ -1,4 +1,4 @@
-package config
+package endpointindex
 
 import (
 	"testing"
