+++ transport/internal/normalize/normalize.go
@@ -1,4 +1,4 @@
-package config
+package normalize
 
 import (
 	"strings"
