+++ transport/internal/normalize/normalize_test.go
@@ -1,4 +1,4 @@
-package config
+package normalize
 
 import (
 	"testing"
