+++ transport/server/http/server.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/Servora-Kit/servora/transport/server/http/cors"
 	"github.com/Servora-Kit/servora/transport/server/http/swagger"
-	sharedendpoint "github.com/Servora-Kit/servora/transport/shared/endpoint"
-	sharedtls "github.com/Servora-Kit/servora/transport/shared/tls"
+	registry "github.com/Servora-Kit/servora/transport/server/internal/registry"
+	tls "github.com/Servora-Kit/servora/transport/internal/tls"
 )
 
 func NewServer(opts ...ServerOption) *khttp.Server {
@@ -41,7 +41,7 @@ func NewServer(opts ...ServerOption) *khttp.Server {
 		if timeout := listen.GetTimeout(); timeout != nil {
 			serverOpts = append(serverOpts, khttp.Timeout(timeout.AsDuration()))
 		}
-		if tlsCfg := sharedtls.MustBuildServerTLS(o.conf.GetTls()); tlsCfg != nil {
+		if tlsCfg := tls.MustBuildServerTLS(o.conf.GetTls()); tlsCfg != nil {
 			serverOpts = append(serverOpts, khttp.TLSConfig(tlsCfg))
 		}
 
@@ -60,7 +60,7 @@ func NewServer(opts ...ServerOption) *khttp.Server {
 		q := url.Values{}
 		q.Set("isSecure", strconv.FormatBool(secure))
 
-		endpoint, err := sharedendpoint.ResolveRegistryEndpoint(sharedendpoint.RegistryEndpointInput{
+		endpoint, err := registry.ResolveRegistryEndpoint(registry.RegistryEndpointInput{
 			Scheme:   scheme,
 			BindAddr: bindAddr,
 			Endpoint: registryEndpoint,
