+++ transport/server/internal/accept/accept_test.go
@@ -1,4 +1,4 @@
-package lifecycle
+package accept
 
 import (
 	"errors"
