+++ transport/server/internal/registry/registry_test.go
@@ -1,4 +1,4 @@
-package endpoint
+package registry
 
 import (
 	"net/url"
